requirements.txt
openai-whisper
ffmpeg-python
srt
